refactor util functions for pheno command into new module
